+++ crates/storage/src/lib.rs
@@ -2,6 +2,7 @@ mod storage;
 mod storage_data;
 mod storage_data_file;
 mod storage_file;
+mod storage_file_try_from;
 mod storage_impl;
 mod storage_impl_drop;
 mod storage_impl_storage;
